+++ coil-base/src/main/java/coil/util/Extensions.kt
@@ -7,6 +7,7 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.drawable.BitmapDrawable
 import android.graphics.drawable.Drawable
+import android.net.Uri
 import android.os.Build.VERSION.SDK_INT
 import android.os.Build.VERSION_CODES.JELLY_BEAN_MR2
 import android.os.Build.VERSION_CODES.KITKAT
@@ -177,6 +178,9 @@ internal fun lazyCallFactory(initializer: () -> Call.Factory): Call.Factory {
     return Call.Factory { lazy.value.newCall(it) } // Intentionally not a method reference.
 }
 
-internal inline fun MimeTypeMap.getMimeTypeFromUrl(url: String): String? {
+internal fun MimeTypeMap.getMimeTypeFromUrl(url: String): String? {
     return getMimeTypeFromExtension(MimeTypeMap.getFileExtensionFromUrl(url))
 }
+
+internal val Uri.firstPathSegment: String?
+    get() = pathSegments.firstOrNull()
